+++ OrganicGLWinLib/ImGuiSliderFloatPanelContainer.h
@@ -11,7 +11,7 @@ class ImGuiSliderFloatPanelContainer
 {
 	public:
 		std::unordered_map<std::string, ImGuiSliderFloatPanel> panels;
-		void insertNewPanel(std::string in_panelName, int in_xOffset, int in_yOffset, int in_xSize, int in_ySize)
+		void insertNewPanel(std::string in_panelName, float in_xOffset, float in_yOffset, float in_xSize, float in_ySize)
 		{
 			ImGuiSliderFloatPanel newPanel(in_panelName, in_xOffset, in_yOffset, in_xSize, in_ySize);
 			panels[in_panelName] = newPanel;
